A better test for ensuring no loop between typeglobs and symbol tables.
Commit e15faf7d09c73a41 in June 2005 removed the reference count loop
between typeglobs and symbol tables. However the test that it used happens
to depend on the internal flag for global destruction not actually being set
until just after the optree for the main program has been freed up, because
the last reference keeping the typeglob alive was held by the main program.
Hence message from destroy that it was searching for actually happened in
perl_destruct(), but before global destruction had been "announced".

This commit provides a more robust test, that doesn't rely on the particular
order of C code within perl_destruct(). It passes on e15faf7d09c73a41, but
fails on e15faf7d09c73a41^.